requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
crispy-bootstrap5==0.7
Django==4.2.1
django-bootstrap-v5==1.0.11
django-countries==7.5.1
django-crispy-forms==2.0
django-filter==23.2
django-widget-tweaks==1.4.12
Pillow==9.5.0
soupsieve==2.4.1
sqlparse==0.4.4
typing-extensions==4.5.0
